+++ internal/token/service/token_service_test.go
@@ -34,6 +34,7 @@ func setupTestTokenService(t *testing.T) (*TokenService, func()) {
 		7*24*time.Hour,
 		redisClient,
 		blacklist,
+		nil,
 		logger,
 	)